+++ lib/l10n/supported_languages.dart
@@ -14,7 +14,11 @@ final Map<Locale, (String, String)> supportedLanguages = {
   const Locale("de", "DE"): ("German (Germany)", "Deutsch (Deutschland)"),
   const Locale("ru", "RU"): ("Russian (Russia)", "Русский (Россия)"),
   const Locale("es", "ES"): ("Spanish (Spain)", "Español (España)"),
+  const Locale("pl", "PL"): ("Polish (Poland)", "Polski (Polska)"),
+  const Locale("be", "BY"): ("Belarusian (Belarus)", "Беларуская (Беларусь)"),
   const Locale("uk", "UA"): ("Ukrainian (Ukraine)", "Українська (Україна)"),
   const Locale("ar"): ("Arabic", "العربية"),
   const Locale("fa", "IR"): ("Persian (Iran)", "فارسی (ایران)"),
+  const Locale("zh", "CN"): ("Chinese (Simplified, China Mainland)", "简体中文 (中国大陆)"),
+  const Locale("zh", "TW"): ("Chinese (Traditional, Taiwan)", "正體中文 (台灣)"),
 };