+++ scripts/translate_missing.dart
@@ -50,10 +50,14 @@ final Map<String, String> filenameToTargetLanguageMapping = {
   "fr_FR.json": "French (France)",
   "fa_IR.json": "Persian (Iran)",
   "it_IT.json": "Italian (Italy)",
+  "pl_PL.json": "Polish (Poland)",
+  "be_BY.json": "Belarusian (Belarus)",
   "mn_MN.json": "Mongolian (Mongolia)",
   "ru_RU.json": "Russian (Russia)",
   "tr_TR.json": "Turkish (Turkey)",
   "uk_UA.json": "Ukrainian (Ukraine)",
+  "zh_CN.json": "Simplified Chinese (China)",
+  "zh_TW.json": "Traditional Chinese (Taiwan)",
 };
 
 Future<void> translateMissingKeys(